requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
dask==2025.5.1
fsspec==2025.5.1
idna==3.10
locket==1.0.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pandas_ta==0.3.14b0
partd==1.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
toolz==1.0.0
tzdata==2025.2
urllib3==2.5.0
